+++ onnxruntime/contrib_ops/cuda/bert/attention.h
@@ -6,7 +6,7 @@
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
 #include "core/providers/cuda/cuda_common.h"
-#include "contrib_ops/cpu/bert/attention.h"
+#include "contrib_ops/cpu/bert/attention_base.h"
 
 namespace onnxruntime {
 namespace contrib {
